resume-parser-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.0.0
pypdf2>=3.0.0
python-docx>=1.0.0
